add first iteration of the AGWB calculation with jax backend
this code was first developed in 2020-2021 by E.Bellini, G.Cusin, I.Michaloliakos, B.Whiting
for usage in Montepython PE on the anisotropies of the astrophysical gravitational wave background.
It relies on the theoretical work by G.Cusin et al on the clustering `C_ell`'s,
the work by G.Cusin et al, A.Jenkins et al. on the shot noise contributions to the AGWB
and the work by D.Alonso et al on the instrumental noise `N_ell`'s.
It was initially developed with CLASS as the cosmological Boltzmann code, and PE with montepython.
Later adapted for PE with Ultranest. Here adapted for PE with` jax_cosmo` as the cosmolopgical backbone and `numpyro` as the PE backbone.

It draws concepts from the Zaldariaga/Seljak work to efficiently compute the integrals over `k` and `z`